// src/app/swagger/models/anprEventsResponse.ts
/**
 * Engine api
 * Engine APIs
 *
 * OpenAPI spec version: 1.0.2
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */import { AnprEvent } from './anprEvent';
import { Links } from './links';
import { Meta } from './meta';


export interface AnprEventsResponse { 
    items?: Array<AnprEvent>;
    links?: Links;
    meta?: Meta;
}