// src/app/swagger/models/job.ts
/**
 * Engine api
 * Engine APIs
 *
 * OpenAPI spec version: 1.0.2
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Links } from './links';

export interface Job { 
    readonly _id?: string;
    /**
     * Job type - anpr, recognition etc. etc.
     */
    jobType?: string;
    /**
     * Job Submission time
     */
    submissionTime?: number;
    videoFileNames?: Array<string>;
    readonly updated?: Date;
    readonly created?: Date;
    readonly etag?: string;
    links?: Links;
}