// src/app/swagger/models/meta.ts
/**
 * Engine api
 * Engine APIs
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Meta { 
    page?: number;
    maxResults?: number;
    total?: number;
}